employee-insight-portal/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
imbalanced-learn>=0.11.0
langchain>=0.0.340
langchain-openai>=0.0.2
chromadb>=0.4.18
openai>=1.6.1
pydantic>=2.5.0
python-multipart>=0.0.6
